+++ src/shims/unix/linux/foreign_items.rs
@@ -4,10 +4,10 @@ use rustc_target::spec::abi::Abi;
 
 use crate::*;
 use shims::foreign_items::EmulateByNameResult;
-use shims::posix::fs::EvalContextExt as _;
-use shims::posix::linux::sync::futex;
-use shims::posix::sync::EvalContextExt as _;
-use shims::posix::thread::EvalContextExt as _;
+use shims::unix::fs::EvalContextExt as _;
+use shims::unix::linux::sync::futex;
+use shims::unix::sync::EvalContextExt as _;
+use shims::unix::thread::EvalContextExt as _;
 
 impl<'mir, 'tcx: 'mir> EvalContextExt<'mir, 'tcx> for crate::MiriEvalContext<'mir, 'tcx> {}
 pub trait EvalContextExt<'mir, 'tcx: 'mir>: crate::MiriEvalContextExt<'mir, 'tcx> {
